+++ rocAL/source/pipeline/master_graph.cpp
@@ -1131,15 +1131,22 @@ std::vector<rocalTensorList *> MasterGraph::create_tf_record_meta_data_reader(co
     return _metadata_output_tensor_list;
 }
 
-std::vector<rocalTensorList *> MasterGraph::create_label_reader(const char *source_path, MetaDataReaderType reader_type) {
+std::vector<rocalTensorList *> MasterGraph::create_label_reader(const char *source_path, MetaDataReaderType reader_type, const char *file_list_path) {
     if (_meta_data_reader)
         THROW("A metadata reader has already been created")
     if (_augmented_meta_data)
         THROW("Metadata can only have a single output")
-
-    MetaDataConfig config(MetaDataType::Label, reader_type, source_path);
+    if (strlen(source_path) == 0 && strlen(file_list_path) == 0)
+        THROW("Either source path or file list path needs to be provided")
+    const char *root_path;
+    if (strlen(file_list_path) == 0) {
+        root_path = source_path;
+    } else {
+        root_path = file_list_path;
+    }
+    MetaDataConfig config(MetaDataType::Label, reader_type, root_path);
     _meta_data_reader = create_meta_data_reader(config, _augmented_meta_data);
-    _meta_data_reader->read_all(source_path);
+    _meta_data_reader->read_all(root_path);
 
     std::vector<size_t> dims = {1};
     auto default_labels_info = TensorInfo(std::move(dims), _mem_type, RocalTensorDataType::INT32);  // Create default labels Info